Refactor(eos_designs): Make WAN policies great again
