+++ ansible_collections/arista/avd/molecule/eos_designs_unit_tests/intended/configs/cv-pathfinder-edge-no-common-path-group.cfg
@@ -111,6 +111,19 @@ router path-selection
       !
       peer dynamic
    !
+   path-group INET id 101
+      ipsec profile CP-PROFILE
+      !
+      local interface Ethernet3
+         stun server-profile INET-cv-pathfinder-pathfinder-Ethernet1 INET-cv-pathfinder-pathfinder-Ethernet3
+      !
+      peer dynamic
+      !
+      peer static router-ip 192.168.144.1
+         name cv-pathfinder-pathfinder
+         ipv4 address 10.7.7.7
+         ipv4 address 10.9.9.9
+   !
    path-group Satellite id 104
       ipsec profile CP-PROFILE
       keepalive interval auto
@@ -120,24 +133,31 @@ router path-selection
       peer dynamic
    !
    load-balance policy LB-DEFAULT-AVT-POLICY-CONTROL-PLANE
+      path-group INET
       path-group Satellite priority 2
    !
    load-balance policy LB-DEFAULT-AVT-POLICY-DEFAULT
+      path-group INET
    !
    load-balance policy LB-DEFAULT-AVT-POLICY-VIDEO
+      path-group INET
    !
    load-balance policy LB-DEFAULT-POLICY-DEFAULT
       path-group AWS
+      path-group INET
       path-group Satellite priority 2
    !
    load-balance policy LB-PROD-AVT-POLICY-DEFAULT
+      path-group INET
    !
    load-balance policy LB-PROD-AVT-POLICY-VIDEO
       loss-rate 42.0
+      path-group INET priority 2
    !
    load-balance policy LB-PROD-AVT-POLICY-VOICE
       hop count lowest
       jitter 42
+      path-group INET priority 2
 !
 spanning-tree mode none
 !
@@ -199,6 +219,14 @@ interface Ethernet1
 !
 interface Ethernet2
    description AWS-1_212
+   no shutdown
+   no switchport
+   flow tracker hardware WAN-FLOW-TRACKER
+   ip address dhcp
+   dhcp client accept default-route
+!
+interface Ethernet3
+   description TEMPORARY FOR PR SAKE
    no shutdown
    no switchport
    flow tracker hardware WAN-FLOW-TRACKER
@@ -286,6 +314,7 @@ application traffic recognition
       42.42.42.0/24
    !
    field-set ipv4 prefix PFX-PATHFINDERS
+      192.168.144.1/32
    !
    field-set l4-port TCP-DEST-2
       666, 777
@@ -354,6 +383,8 @@ router bgp 65000
    neighbor 172.17.0.2 peer group IPv4-UNDERLAY-PEERS
    neighbor 172.17.0.2 remote-as 65199
    neighbor 172.17.0.2 description site-ha-disabled-leaf_Ethernet2
+   neighbor 192.168.144.1 peer group WAN-OVERLAY-PEERS
+   neighbor 192.168.144.1 description cv-pathfinder-pathfinder
    redistribute connected route-map RM-CONN-2-BGP
    !
    address-family evpn
@@ -427,5 +458,14 @@ management security
       tls versions 1.2
       trust certificate aristaDeviceCertProvisionerDefaultRootCA.crt
       certificate profileA.crt key profileA.key
+!
+stun
+   client
+      server-profile INET-cv-pathfinder-pathfinder-Ethernet1
+         ip address 10.7.7.7
+         ssl profile profileA
+      server-profile INET-cv-pathfinder-pathfinder-Ethernet3
+         ip address 10.9.9.9
+         ssl profile profileA
 !
 end
