+++ ansible_collections/arista/avd/molecule/eos_designs_unit_tests/intended/structured_configs/cv-pathfinder-edge-no-common-path-group.yml
@@ -41,6 +41,10 @@ router_bgp:
     remote_as: '65199'
     peer: site-ha-disabled-leaf
     description: site-ha-disabled-leaf_Ethernet2
+  - ip_address: 192.168.144.1
+    peer_group: WAN-OVERLAY-PEERS
+    peer: cv-pathfinder-pathfinder
+    description: cv-pathfinder-pathfinder
   vrfs:
   - name: IT
     router_id: 192.168.42.2
@@ -218,6 +222,15 @@ ethernet_interfaces:
   shutdown: false
   type: routed
   description: AWS-1_212
+  dhcp_client_accept_default_route: true
+  flow_tracker:
+    hardware: WAN-FLOW-TRACKER
+- name: Ethernet3
+  peer_type: l3_interface
+  ip_address: dhcp
+  shutdown: false
+  type: routed
+  description: TEMPORARY FOR PR SAKE
   dhcp_client_accept_default_route: true
   flow_tracker:
     hardware: WAN-FLOW-TRACKER
@@ -463,26 +476,66 @@ router_path_selection:
     dynamic_peers:
       enabled: true
     ipsec_profile: CP-PROFILE
+  - name: INET
+    id: 101
+    local_interfaces:
+    - name: Ethernet3
+      stun:
+        server_profiles:
+        - INET-cv-pathfinder-pathfinder-Ethernet1
+        - INET-cv-pathfinder-pathfinder-Ethernet3
+    dynamic_peers:
+      enabled: true
+    static_peers:
+    - router_ip: 192.168.144.1
+      name: cv-pathfinder-pathfinder
+      ipv4_addresses:
+      - 10.7.7.7
+      - 10.9.9.9
+    ipsec_profile: CP-PROFILE
   load_balance_policies:
   - name: LB-DEFAULT-AVT-POLICY-CONTROL-PLANE
     path_groups:
+    - name: INET
     - name: Satellite
       priority: 2
   - name: LB-DEFAULT-AVT-POLICY-VIDEO
+    path_groups:
+    - name: INET
   - name: LB-DEFAULT-AVT-POLICY-DEFAULT
+    path_groups:
+    - name: INET
   - name: LB-PROD-AVT-POLICY-VOICE
+    path_groups:
+    - name: INET
+      priority: 2
     jitter: 42
     lowest_hop_count: true
   - name: LB-PROD-AVT-POLICY-VIDEO
+    path_groups:
+    - name: INET
+      priority: 2
     loss_rate: '42.0'
   - name: LB-PROD-AVT-POLICY-DEFAULT
+    path_groups:
+    - name: INET
   - name: LB-DEFAULT-POLICY-DEFAULT
     path_groups:
+    - name: INET
     - name: Satellite
       priority: 2
     - name: AWS
 router_traffic_engineering:
   enabled: true
+stun:
+  client:
+    server_profiles:
+    - name: INET-cv-pathfinder-pathfinder-Ethernet1
+      ip_address: 10.7.7.7
+      ssl_profile: profileA
+    - name: INET-cv-pathfinder-pathfinder-Ethernet3
+      ip_address: 10.9.9.9
+      ssl_profile: profileA
 application_traffic_recognition:
   application_profiles:
   - name: VIDEO
@@ -533,6 +586,8 @@ application_traffic_recognition:
       prefix_values:
       - 6.6.6.0/24
     - name: PFX-PATHFINDERS
+      prefix_values:
+      - 192.168.144.1/32
 dps_interfaces:
 - name: Dps1
   description: DPS Interface
@@ -599,6 +654,14 @@ metadata:
         value: AWS-1
       - name: Circuit
         value: '212'
+    - interface: Ethernet3
+      tags:
+      - name: Type
+        value: wan
+      - name: Carrier
+        value: ATT
+      - name: Circuit
+        value: '404'
   cv_pathfinder:
     role: edge
     vtep_ip: 192.168.142.2
@@ -614,3 +677,9 @@ metadata:
       carrier: AWS-1
       circuit_id: '212'
       pathgroup: AWS
+    - name: Ethernet3
+      carrier: ATT
+      circuit_id: '404'
+      pathgroup: INET
+    pathfinders:
+    - vtep_ip: 192.168.144.1
