+++ ansible_collections/arista/avd/molecule/eos_designs_unit_tests/intended/structured_configs/cv-pathfinder-pathfinder1.yml
@@ -356,7 +356,6 @@ router_path_selection:
   load_balance_policies:
   - name: LB-DEFAULT-AVT-POLICY-CONTROL-PLANE
     path_groups:
-    - name: LAN_HA
     - name: MPLS
     - name: INET
     - name: Equinix
@@ -364,57 +363,57 @@ router_path_selection:
     - name: Satellite
       priority: 2
     - name: AWS
+    - name: LAN_HA
   - name: LB-DEFAULT-AVT-POLICY-VIDEO
     path_groups:
-    - name: LAN_HA
     - name: MPLS
     - name: INET
     - name: Equinix
+    - name: LAN_HA
   - name: LB-DEFAULT-AVT-POLICY-DEFAULT
     path_groups:
-    - name: LAN_HA
     - name: INET
     - name: Equinix
       priority: 2
     - name: MPLS
       priority: 4223
+    - name: LAN_HA
   - name: LB-PROD-AVT-POLICY-VOICE
     path_groups:
-    - name: LAN_HA
     - name: MPLS
     - name: INET
       priority: 2
+    - name: LAN_HA
     jitter: 42
     lowest_hop_count: true
   - name: LB-PROD-AVT-POLICY-VIDEO
     path_groups:
-    - name: LAN_HA
     - name: MPLS
     - name: LTE
     - name: INET
       priority: 2
+    - name: LAN_HA
     loss_rate: '42.0'
   - name: LB-PROD-AVT-POLICY-DEFAULT
     path_groups:
-    - name: LAN_HA
     - name: INET
     - name: MPLS
       priority: 2
+    - name: LAN_HA
   - name: LB-CUSTOM-VOICE-PROFILE-NAME
     path_groups:
-    - name: LAN_HA
     - name: MPLS
     - name: INET
       priority: 2
+    - name: LAN_HA
   - name: LB-TRANSIT-AVT-POLICY-DEFAULT
     path_groups:
-    - name: LAN_HA
     - name: INET
     - name: MPLS
       priority: 2
+    - name: LAN_HA
   - name: LB-DEFAULT-POLICY-DEFAULT
     path_groups:
-    - name: LAN_HA
     - name: MPLS
     - name: INET
     - name: Equinix
@@ -422,6 +421,7 @@ router_path_selection:
     - name: Satellite
       priority: 2
     - name: AWS
+    - name: LAN_HA
 router_traffic_engineering:
   enabled: true
 stun:
@@ -576,8 +576,6 @@ metadata:
       - id: 254
         name: DEFAULT-AVT-POLICY-CONTROL-PLANE
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: MPLS
           preference: preferred
         - name: INET
@@ -588,28 +586,30 @@ metadata:
           preference: alternate
         - name: AWS
           preference: preferred
+        - name: LAN_HA
+          preference: preferred
       - id: 3
         name: DEFAULT-AVT-POLICY-VIDEO
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: MPLS
           preference: preferred
         - name: INET
           preference: preferred
         - name: Equinix
+          preference: preferred
+        - name: LAN_HA
           preference: preferred
       - id: 1
         name: DEFAULT-AVT-POLICY-DEFAULT
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: INET
           preference: preferred
         - name: Equinix
           preference: alternate
         - name: MPLS
           preference: alternate
+        - name: LAN_HA
+          preference: preferred
     - name: PROD
       vni: 142
       avts:
@@ -618,86 +618,84 @@ metadata:
         id: 2
         name: PROD-AVT-POLICY-VOICE
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: MPLS
           preference: preferred
         - name: INET
           preference: alternate
+        - name: LAN_HA
+          preference: preferred
       - id: 4
         name: PROD-AVT-POLICY-VIDEO
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: MPLS
           preference: preferred
         - name: LTE
           preference: preferred
         - name: INET
           preference: alternate
+        - name: LAN_HA
+          preference: preferred
       - id: 1
         name: PROD-AVT-POLICY-DEFAULT
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: INET
           preference: preferred
         - name: MPLS
           preference: alternate
+        - name: LAN_HA
+          preference: preferred
     - name: IT
       vni: 1000
       avts:
       - id: 3
         name: DEFAULT-AVT-POLICY-VIDEO
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: MPLS
           preference: preferred
         - name: INET
           preference: preferred
         - name: Equinix
+          preference: preferred
+        - name: LAN_HA
           preference: preferred
       - id: 1
         name: DEFAULT-AVT-POLICY-DEFAULT
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: INET
           preference: preferred
         - name: Equinix
           preference: alternate
         - name: MPLS
           preference: alternate
+        - name: LAN_HA
+          preference: preferred
     - name: TRANSIT
       vni: 66
       avts:
       - id: 42
         name: CUSTOM-VOICE-PROFILE-NAME
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: MPLS
           preference: preferred
         - name: INET
           preference: alternate
+        - name: LAN_HA
+          preference: preferred
       - id: 1
         name: TRANSIT-AVT-POLICY-DEFAULT
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: INET
           preference: preferred
         - name: MPLS
           preference: alternate
+        - name: LAN_HA
+          preference: preferred
     - name: ATTRACTED-VRF-FROM-UPLINK
       vni: 666
       avts:
       - id: 1
         name: DEFAULT-POLICY-DEFAULT
         pathgroups:
-        - name: LAN_HA
-          preference: preferred
         - name: MPLS
           preference: preferred
         - name: INET
@@ -707,4 +705,6 @@ metadata:
         - name: Satellite
           preference: alternate
         - name: AWS
+          preference: preferred
+        - name: LAN_HA
           preference: preferred
